refactor: store/utils in ts
